requirements.txt
accelerate==1.3.0
beautifulsoup4==4.12.3
jupyter==1.1.1
lxml==5.3.0
matplotlib==3.10.0
numpy==2.2.2
optimum-quanto==0.2.6
pandas==2.2.3
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
